Move sorted_neighbors to util and update docstring
