+++ indra/explanation/pathfinding/util.py
@@ -57,3 +57,47 @@ def signed_nodes_to_signed_edge(source, target):
         logger.warning('Error translating signed nodes to signed edge using '
                        '(%s, %s)' % (source, target))
         return None, None, None
+
+
+def get_sorted_neighbors(G, node, reverse, g_edges):
+    """Sort the returned neighbors in descending order by belief
+
+    Parameters
+    ----------
+    G : nx.DiGraph
+        A networkx DiGraph
+    node : str|int
+        A valid networkx node name
+    reverse : bool
+        Indicates direction of search. Neighbors are either successors
+        (downstream search) or predecessors (reverse search).
+    g_edges : nx.classes.reportviews.OutMultiEdgeView|OutEdgeView
+        The edges graph property to look up edges and their data from.
+        Typically this is G.edges.
+    """
+    neighbors = G.predecessors(node) if reverse else G.successors(node)
+    # Check signed node
+    if isinstance(node, tuple):
+        if reverse:
+            return sorted(
+                neighbors,
+                key=lambda n:
+                    g_edges[signed_nodes_to_signed_edge(n, node)]['belief'],
+                reverse=True
+            )
+        else:
+            return sorted(
+                neighbors,
+                key=lambda n:
+                    g_edges[signed_nodes_to_signed_edge(node, n)]['belief'],
+                reverse=True)
+
+    else:
+        if reverse:
+            return sorted(neighbors,
+                          key=lambda n: g_edges[(n, node)]['belief'],
+                          reverse=True)
+        else:
+            return sorted(neighbors,
+                          key=lambda n: g_edges[(node, n)]['belief'],
+                          reverse=True)
